+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - (Bugfix) Propagate Gateway Idle Timeout
 - (Feature) Update Authz V1 Definitions
 - (Feature) (Platform) Add Sharding to the Inventory Endpoint
+- (Feature) Emit Default values in Inventory
 
 ## [1.2.46](https://github.com/arangodb/kube-arangodb/tree/1.2.46) (2025-02-24)
 - (Bugfix) Clean Phase change properly during upgrade